Move the ChannelReporter, Reporter implementation to a submodule of reporter
This logically splits out an implementation detail from the primary trait api.
